+++ src/asp/IsisIO/PolyEquation.h
@@ -36,8 +36,8 @@ namespace asp {
     }
     std::string type() const {  return "PolyEquation"; }
 
-    unsigned size() const { return m_x_coeff.size()+m_y_coeff.size()+m_z_coeff.size(); }
-    double& operator[]( unsigned const& n );
+    size_t size() const { return m_x_coeff.size()+m_y_coeff.size()+m_z_coeff.size(); }
+    double& operator[]( size_t const& n );
 
     void write( std::ofstream &f );
     void read( std::ifstream &f );
